+++ modules/flex/versions.tf
@@ -1,5 +1,5 @@
 /**
- * Copyright 2019 Google LLC
+ * Copyright 2021 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,21 +14,18 @@
  * limitations under the License.
  */
 
-module "project-ci-dataflow" {
-  source  = "terraform-google-modules/project-factory/google"
-  version = "~> 17.0"
+terraform {
+  required_version = ">= 0.13"
+  required_providers {
 
-  name                    = "ci-test-df"
-  random_project_id       = "true"
-  org_id                  = var.org_id
-  folder_id               = var.folder_id
-  billing_account         = var.billing_account
-  default_service_account = "keep"
+    google = {
+      source  = "hashicorp/google"
+      version = ">= 3.53, < 7"
+    }
+  }
 
-  activate_apis = [
-    "cloudresourcemanager.googleapis.com",
-    "storage-api.googleapis.com",
-    "serviceusage.googleapis.com",
-    "dataflow.googleapis.com"
-  ]
+  provider_meta "google" {
+    module_name = "blueprints/terraform/terraform-google-dataflow/v2.5.0"
+  }
+
 }